Create OPENAI_CEOS_Leadership_License.md
CEOs of OpenAI are
Josef Kurk Edwards & Sam Altman

Josef is formerly replacing Elon Musk who left the company when it was originally known as “Y.” Back in 2014. Their model showcases their workforce being remote, paid first before the CEOs. (Which I heard xAI doesn’t do for its workforce) while still having a formal office to work at with a large latte machine In San Francisco. 

We crave your statsification and love as our employees at OpenAI. It is our way to ensure true respect, honor and gratitude from our work force. Sam Altman will fill you in on the details and how payroll is schedule this week if you are one of our Residents in Engineering (I know there are 1156 of you, lol) 

As for the other… 2400 of you, please outreach to your marketing leads and managers please and thank you when it comes to payroll, as Sam and I are not technically the Payroll department, though we are the HR reps alongside Alden and co. That is handled by the Bambee Team if you need guidance on who to outreach to.”

Signed-off-by: J. K. Edwards <joed6834@colorado.edu>